requirements.txt
opencv-python
numpy
PyYAML
pytesseract
